refactor: move into fmgc
