+++ .gitignore
@@ -7,7 +7,7 @@
 /A32NX/layout.json
 /A32NX/manifest.json
 /A32NX/build_info.json
-/A32NX/html_ui/JS/flightPlan/bundle.js
+/A32NX/html_ui/JS/fmgc/bundle.js
 /A32NX/html_ui/Pages/VCockpit/Instruments/generated
 /A32NX/ModelBehaviorDefs/A32NX/generated
 !/docs/**
