+++ A32NX/html_ui/Pages/VCockpit/Instruments/A320_Neo/Map/A32NX_MapInstrument.html
@@ -44,7 +44,7 @@
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/Approach.js"></script>
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/Runway.js"></script>
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/GeoCalc.js"></script>
-<script type="text/html" import-script="/JS/flightPlan/bundle.js"></script>
+<script type="text/html" import-script="/JS/fmgc/bundle.js"></script>
 
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/Map/Canvas/CanvasUtils.js"></script>
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/Map/Cities/City.js"></script>
