+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/Clock/A320_Neo_Clock.html
@@ -40,7 +40,7 @@
 
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/Shared/FlightElements/GeoCalc.js"></script>
 
-<script type="text/html" import-script="/JS/flightPlan/bundle.js"></script>
+<script type="text/html" import-script="/JS/fmgc/bundle.js"></script>
 
 <script type="text/html" import-script="/Pages/VCockpit/Instruments/NavSystems/A320_Neo/A32NX_NavSystem.js"></script>
