+++ package.json
@@ -8,7 +8,7 @@
     "build:instruments": "rollup -c src/instruments/rollup.config.js",
     "postinstall": "husky install",
     "serve:efb": "cd src/instruments/src/EFB/ && rollup -c ./rollup.development.config.js",
-    "build:flightplan": "rollup -c src/flightplan/rollup.config.js"
+    "build:fmgc": "rollup -c src/fmgc/rollup.config.js"
   },
   "lint-staged": {
     "*.{js,mjs,jsx}": "eslint --cache --fix",
