--- src/flightplan/src/types/fstypes/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "fstypes",
-  "version": "1.0.0",
-  "typings": "main.d.ts"
-}
