+++ src/fmgc/src/flightplanning/FixNamingScheme.ts
