+++ src/fmgc/src/flightplanning/FlightPlanAsoboSync.ts
