+++ src/fmgc/src/flightplanning/FlightPlanManager.ts
