+++ src/fmgc/src/flightplanning/FlightPlanSegment.ts
