+++ src/fmgc/src/flightplanning/GPS.ts
