+++ src/fmgc/src/flightplanning/WaypointBuilder.ts
