+++ src/fmgc/src/flightplanning/WorldMagneticModel.ts
