+++ src/fmgc/src/wtsdk.ts
@@ -1,0 +1,28 @@
+/*
+ * A32NX
+ * Copyright (C) 2020-2021 FlyByWire Simulations and its contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { DirectTo } from './flightplanning/DirectTo';
+import { FlightPlanManager } from './flightplanning/FlightPlanManager';
+import { FlightPlanSegment } from './flightplanning/FlightPlanSegment';
+import { GPS } from './flightplanning/GPS';
+import { LegsProcedure } from './flightplanning/LegsProcedure';
+import { ManagedFlightPlan } from './flightplanning/ManagedFlightPlan';
+import { ProcedureDetails } from './flightplanning/ProcedureDetails';
+import { RawDataMapper } from './flightplanning/RawDataMapper';
+
+export { DirectTo, FlightPlanManager, FlightPlanSegment, GPS, LegsProcedure, ManagedFlightPlan, ProcedureDetails, RawDataMapper }
